# frontend/person-blog-app/packages/blog/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Gateway.API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "/video"
    }
  ],
  "paths": {
...
    "/api/ConferenceRoom/createConferenceToPost": {
      "post": {
        "tags": [
          "ConferenceRoom"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ConferenceRoom/joinLink": {
      "get": {
        "tags": [
          "ConferenceRoom"
        ],
        "parameters": [
          {
            "name": "roomId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ConferenceRoom/join": {
      "get": {
        "tags": [
          "ConferenceRoom"
        ],
        "parameters": [
          {
            "name": "roomId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}